Move jobs resume methods to azcopy
