+++ cmd/constants.go
@@ -56,14 +56,5 @@ const (
 	base10Mega = 1000 * 1000
 )
 
-// credentials related consts
-const (
-	oauthLoginSessionCacheKeyName     = "AzCopyOAuthTokenCache"
-	oauthLoginSessionCacheServiceName = "AzCopyV10"
-	oauthLoginSessionCacheAccountName = "AzCopyOAuthTokenCache"
-	trustedSuffixesNameAAD            = "trusted-microsoft-suffixes"
-	trustedSuffixesAAD                = "*.core.windows.net;*.core.chinacloudapi.cn;*.core.cloudapi.de;*.core.usgovcloudapi.net;*.storage.azure.net"
-)
-
 const MAX_SYMLINKS_TO_FOLLOW = 40
 const NumOfFilesPerDispatchJobPart = 10000
