+++ cmd/zc_traverser_local_test.go
@@ -1,9 +1,11 @@
 package cmd
 
 import (
-	"github.com/stretchr/testify/assert"
 	"os"
 	"testing"
+
+	"github.com/Azure/azure-storage-azcopy/v10/azcopy"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestCleanLocalPath(t *testing.T) {
@@ -18,7 +20,7 @@ func TestCleanLocalPath(t *testing.T) {
 	}
 
 	for orig, expected := range testCases {
-		a.Equal(expected, cleanLocalPath(orig))
+		a.Equal(expected, azcopy.CleanLocalPath(orig))
 	}
 }
 
@@ -45,6 +47,6 @@ func TestCleanLocalPathForWindows(t *testing.T) {
 	}
 
 	for orig, expected := range testCases {
-		a.Equal(expected, cleanLocalPath(orig))
+		a.Equal(expected, azcopy.CleanLocalPath(orig))
 	}
 }
